requirements.txt
langgraph==0.2.34
langchain==0.3.7
langchain-community==0.3.5
huggingface-hub==0.26.2
python-dotenv==1.0.0
pypdf2==3.0.1
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3
pydantic>=2.7.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pytesseract 
numpy~=1.24.3
streamlit~=1.52.2